gen-llm/pyproject.toml
[tool.poetry]
name = "auto-sim-hub"
version = "0.1.0"
description = "AutoSimHub General LLM"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.51.0"
fastapi = "^0.115.0"
python-dotenv = "^1.0.1"
open-interpreter = "^0.3.13"
simpy = "^4.1.1"
pandas = "^2.2.3"
graphviz = "^0.20.3"
numpy = "^2.1.2"
gradio = "^4.44.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
